seperated out components
