+++ src/views/Token/components/FailedClaim/index.tsx
@@ -6,19 +6,15 @@ import { useTranslation } from 'react-i18next'
 import { CardBody } from 'src/components/CardBody'
 import { CardText } from 'src/components/CardText'
 import { Flex } from 'src/components/Flex'
-import { Spinner } from 'src/components/Spinner'
 import { Card } from 'src/components/Card'
 
-export const VerifyState = () => {
+export const FailedClaim = () => {
   const [t] = useTranslation()
   return (
     <Card>
       <CardBody height="100%" textAlign="center">
         <Flex height="100%" width="100%" justifyContent="center" flexDirection="column">
-          <Flex justifyContent="center">
-            <Spinner />
-          </Flex>
-          <CardText>{t('texts.verifyTransaction')}</CardText>
+          <CardText>{t('texts.failedTransaction')}</CardText>
         </Flex>
       </CardBody>
     </Card>
