+++ src/views/Token/components/TokenClaim/index.tsx
@@ -1,11 +1,11 @@
 // Externals
 import React, { useEffect, useState } from 'react'
-import SVG from 'react-inlinesvg'
 import { useWeb3React } from '@web3-react/core'
 import { useTranslation } from 'react-i18next'
-import styled from 'styled-components'
 import { Signer } from '@ethersproject/abstract-signer'
 import { ethers } from 'ethers'
+import styled from 'styled-components'
+import { space, SpaceProps } from 'styled-system'
 
 // Components
 import { Card } from 'src/components/Card'
@@ -17,21 +17,19 @@ import { TokenIconFigure } from 'src/components/TokenIconFigure'
 import { Button } from 'src/components/Button'
 import { CardTitle } from 'src/components/CardTitle'
 import { VerifyState } from 'src/views/Token/components/VerifyClaim'
+import { SuccessfulClaim } from 'src/views/Token/components/SuccessfulClaim'
 
 // interface
 import { FixedPriceSalePurchase } from 'src/interfaces/Sale'
 
 // Svg
 import noToken from 'src/assets/svg/no-token-image.svg'
-import check from 'src/assets/svg/Check-Icon.svg'
-import link from 'src/assets/svg/External-Link.svg'
 
 // hooks
 import { useWindowSize } from 'src/hooks/useWindowSize'
 
 // Theme
 import { theme } from 'src/styles/theme'
-import { space, SpaceProps } from 'styled-system'
 
 // Mesa Utils
 import { formatBigInt } from 'src/utils/Defaults'
@@ -40,30 +38,9 @@ import { formatBigInt } from 'src/utils/Defaults'
 import { FixedPriceSaleTemplate__factory } from 'src/contracts'
 
 // interfaces
-interface TokenClaimProps {
+export interface TokenClaimProps {
   purchase: FixedPriceSalePurchase
 }
-
-const Circle = styled.div({
-  height: '45px',
-  background: 'rgba(75, 158, 152, 0.35)',
-  width: '45px',
-  borderRadius: '50%',
-})
-
-const Link = styled.p<SpaceProps>(
-  {
-    color: '#304FFE',
-    cursor: 'pointer',
-    hover: 'underline',
-  },
-  space
-)
-
-const StyledSVG = styled(SVG)(props => ({
-  fill: props.color,
-}))
-
 const Icon = styled.img<SpaceProps>(
   {
     height: '32px',
@@ -81,15 +58,15 @@ export const TokenClaim = ({ purchase: { sale, amount } }: TokenClaimProps) => {
   const claimTokens = async (saleId: string, signer: Signer) => {
     await FixedPriceSaleTemplate__factory.connect(saleId, signer)
       .claimToken()
-      .then(tx => {
+      .then((tx: any) => {
         setClaim('verify')
         tx.wait(1)
       })
-      .then(receipt => {
+      .then((receipt: any) => {
         console.log(receipt)
         setClaim('claimed')
       })
-      .catch(error => {
+      .catch((error: Error) => {
         console.log(error)
         setClaim('failed')
       })
@@ -101,66 +78,40 @@ export const TokenClaim = ({ purchase: { sale, amount } }: TokenClaimProps) => {
     if (!chainId || !library || !account) {
       return
     }
-    // An example Provider
     const provider = new ethers.providers.Web3Provider(library)
-    // An example Signer
     signer = provider.getSigner(0)
   }, [chainId, library, account])
 
-  const preDecimalAmount = amount.toString().split('\\.')[0]
-  const postDecimalAmount = amount.toString().split('\\.')[1]
-  const claimState = (
-    <CardBody padding={theme.space[3]}>
-      <Flex margin="0 0 16px 0">
-        <TokenIconFigure>
-          <Icon src={sale?.tokenOut.icon || noToken} />
-        </TokenIconFigure>
-        <CardTitle fontWeight={500}>Claim {sale?.tokenOut.name}</CardTitle>
-      </Flex>
-      <Divider />
-      <Flex flexDirection="column" justifyContent="space-evenly">
-        <Flex justifyContent="space-between">
-          <CardText color="grey">{t('texts.unclaimed')}</CardText>
-          <Flex>
-            <CardText>{preDecimalAmount}</CardText>
-            <CardText color="grey">{postDecimalAmount}</CardText>
-            <CardText>&nbsp;{sale?.tokenOut.name}</CardText>
+  const preDecimalAmount = formatBigInt(amount, sale?.tokenOut.decimals).toString().split('\\.')[0]
+  const postDecimalAmount = formatBigInt(amount, sale?.tokenOut.decimals).toString().split('\\.')[1]
+  return (
+    <Card>
+      <CardBody padding={theme.space[3]}>
+        <Flex margin="0 0 16px 0">
+          <TokenIconFigure>
+            <Icon src={sale?.tokenOut.icon || noToken} />
+          </TokenIconFigure>
+          <CardTitle fontWeight={500}>Claim {sale?.tokenOut.name}</CardTitle>
+        </Flex>
+        <Divider />
+        <Flex flexDirection="column" justifyContent="space-evenly">
+          <Flex justifyContent="space-between">
+            <CardText color="grey">{t('texts.unclaimed')}</CardText>
+            <Flex>
+              <CardText>{preDecimalAmount}</CardText>
+              <CardText color="grey">{postDecimalAmount}</CardText>
+              <CardText>&nbsp;{sale?.tokenOut.name}</CardText>
+            </Flex>
           </Flex>
+          <Flex justifyContent="space-between">
+            <CardText color="grey">{t('texts.currentPrice')}</CardText>
+            <CardText>{`${sale?.tokenPrice} ${sale?.tokenIn.name}`}</CardText>
+          </Flex>
+          <Button onClick={() => claimTokens(sale!.id, signer)} width="90%">
+            {isMobile ? t('buttons.shortClaim') : t('buttons.claimTokens')}
+          </Button>
         </Flex>
-        <Flex justifyContent="space-between">
-          <CardText color="grey">{t('texts.currentPrice')}</CardText>
-          <CardText>{`${sale?.tokenPrice} ${sale?.tokenIn.name}`}</CardText>
-        </Flex>
-        <Button onClick={() => claimTokens(sale!.id, signer)} width="90%">
-          {isMobile ? t('buttons.shortClaim') : t('buttons.claimTokens')}
-        </Button>
-      </Flex>
-    </CardBody>
+      </CardBody>
+    </Card>
   )
-
-  const claimedState = (
-    <CardBody height="100%" textAlign="center">
-      <Flex flexDirection="column" height="100%">
-        <Flex justifyContent="center">
-          <Circle>
-            <Icon marginTop="8px" src={check} />
-          </Circle>
-        </Flex>
-        <CardTitle fontWeight={500}>{t('texts.claimSuccessful')}</CardTitle>
-        <CardText color="grey">
-          {' '}
-          `${amount} ${sale?.tokenOut.name} has been sent to your address.`
-        </CardText>
-        <Link marginTop="24px">
-          See this transaction on block explorer
-          <StyledSVG src={link} color="#304FFE" />
-        </Link>
-        <Button variant="secondary" width="90%">
-          {t('buttons.done')}
-        </Button>
-      </Flex>
-    </CardBody>
-  )
-
-  return <Card>{claim == 'verify' ? <VerifyState /> : claim == 'claimed' ? claimedState : claimState}</Card>
 }
